requirements.txt
xmltodict
avro
pyyaml
msgpack